feat: Remove account_microfrontend_url from SiteConfiguration Model
As part of the removal of IDV, references to IDV have been removed from ecommerce. Please see https://github.com/edx/ecommerce/pull/3568. The associated JIRA ticket is https://openedx.atlassian.net/browse/MST-856.

This code change is a follow up to that pull request and contains a migration to remove the account_microfrontend_url from the SiteConfiguration model, as it will no longer be in use.

The associated JIRA ticket for this work is https://openedx.atlassian.net/browse/MST-1158.